move key scheduling to seperate file
