+++ src/lib.rs
@@ -30,6 +30,7 @@
 mod tls13formats;
 mod tls13handshake;
 mod tls13record;
+mod tls13keyscheduler;
 
 pub mod server;
 pub mod tls13api;
